perte de pv des pieces WIP

+++ Assets/Scripts/SO_Scripts/pieces/Board/BoardPiece.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 [Serializable]
 public class BoardPiece
 {
-    public float healthPoint;
-    public float shield;
+    public int healthPoint;
+    public int shield;
 
     public SoPieces soPieces;
     public PieceAnimations piecePersonality;
